fix(files_sharing): Improve recursion in `onNewLinShare`
These changes fixes the issue of having the refresh the UI after share creation, as the share is not
immediately removed from the UI list.

Important changes

- The basis of checking wether a password/expire date is no longer based on config values alone
because the config is not expected in a share creation circle. Hence we check the configs and check
if the share object (this.share) has the expected values set. This way, once the required properties
are set, code control does not enter the block meant to handle the setting of required properties
unneccessarily.

Signed-off-by: fenn-cs <fenn25.fn@gmail.com>